renamed post to aux per our phone call with TiD, cleaned up errant dig.py, added templates for linux and windows.
