+++ threat/modules/aux/encodeall.py
