+++ threat/modules/aux/honeypot.py
