--- threat/templates/template.txt
@@ -1,1 +1,0 @@
-
